# KAFKA-9685: PT2, avoid unnecessary set creation in ACL matching