@lbradstreet 2020-03-29
Don't look at the iterators, nothing changed there. testAuthorizer is 27% to 62% of the cost. You could argue that it's already fast enough as it is, but for metadata requests I think it may be called thousands of times.